# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
    "/api/user/info" : {
      "put" : {
        "summary" : "更改用户信息",
        "description" : "更改用户信息",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "password", "username" ],
                "type" : "object",
                "properties" : {
                  "username" : {
                    "type" : "string",
                    "description" : "",
                    "example" : "test"
                  },
                  "password" : {
                    "type" : "string",
                    "description" : "",
                    "example" : "123456"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "username" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "email" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "avatarUrl" : {
                          "type" : "string",
                          "description" : "",
                          "nullable" : true
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/getRole" : {
      "get" : {
        "summary" : "获取用户权限信息",
        "description" : "获取用户权限信息",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "role" : {
                          "type" : "integer",
                          "description" : ""
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/upload" : {
      "post" : {
        "summary" : "用于上传头像",
        "description" : "用于上传头像",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "required" : [ "file" ],
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "description" : "",
                    "format" : "binary"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{id}" : {
      "delete" : {
        "summary" : "删除用户",
        "description" : "删除用户",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/articles" : {
      "post" : {
        "summary" : "创建文章",
        "description" : "创建文章",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "title" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "content" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/articles/upload" : {
      "post" : {
        "summary" : "通过文件上传文章",
        "description" : "通过文件上传文章",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "required" : [ "file" ],
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "description" : "",
                    "format" : "binary"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/article/{articleId}" : {
      "put" : {
        "summary" : "更新文章",
        "description" : "更新文章",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "title" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "content" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/comments/{commentId}" : {
      "put" : {
        "summary" : "删除评论",
        "description" : "删除评论",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "path",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/comments/approve/{commentId}" : {
      "put" : {
        "summary" : "审核评论状态",
        "description" : "审核评论状态",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "path",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : ""
                    },
                    "message" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "string",
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}